Move collect test to its own file
